+++ zcash_primitives/src/transaction/components/sapling.rs
@@ -99,6 +99,33 @@ impl MapAuth<Authorized, Authorized> for () {
     }
 }
 
+/// A helper for implementing `MapAuth` with a set of closures.
+impl<A, B, F, G, H, I> MapAuth<A, B> for (F, G, H, I)
+where
+    A: Authorization,
+    B: Authorization,
+    F: FnMut(A::SpendProof) -> B::SpendProof,
+    G: FnMut(A::OutputProof) -> B::OutputProof,
+    H: FnMut(A::AuthSig) -> B::AuthSig,
+    I: FnMut(A) -> B,
+{
+    fn map_spend_proof(&mut self, p: A::SpendProof) -> B::SpendProof {
+        self.0(p)
+    }
+
+    fn map_output_proof(&mut self, p: A::OutputProof) -> B::OutputProof {
+        self.1(p)
+    }
+
+    fn map_auth_sig(&mut self, s: A::AuthSig) -> B::AuthSig {
+        self.2(s)
+    }
+
+    fn map_authorization(&mut self, a: A) -> B {
+        self.3(a)
+    }
+}
+
 /// A fallible map from one bundle authorization to another.
 ///
 /// For use with [`Bundle::try_map_authorization`].
@@ -108,6 +135,35 @@ pub trait TryMapAuth<A: Authorization, B: Authorization> {
     fn try_map_output_proof(&mut self, p: A::OutputProof) -> Result<B::OutputProof, Self::Error>;
     fn try_map_auth_sig(&mut self, s: A::AuthSig) -> Result<B::AuthSig, Self::Error>;
     fn try_map_authorization(&mut self, a: A) -> Result<B, Self::Error>;
+}
+
+/// A helper for implementing `TryMapAuth` with a set of closures.
+impl<A, B, E, F, G, H, I> TryMapAuth<A, B> for (F, G, H, I)
+where
+    A: Authorization,
+    B: Authorization,
+    F: FnMut(A::SpendProof) -> Result<B::SpendProof, E>,
+    G: FnMut(A::OutputProof) -> Result<B::OutputProof, E>,
+    H: FnMut(A::AuthSig) -> Result<B::AuthSig, E>,
+    I: FnMut(A) -> Result<B, E>,
+{
+    type Error = E;
+
+    fn try_map_spend_proof(&mut self, p: A::SpendProof) -> Result<B::SpendProof, Self::Error> {
+        self.0(p)
+    }
+
+    fn try_map_output_proof(&mut self, p: A::OutputProof) -> Result<B::OutputProof, Self::Error> {
+        self.1(p)
+    }
+
+    fn try_map_auth_sig(&mut self, s: A::AuthSig) -> Result<B::AuthSig, Self::Error> {
+        self.2(s)
+    }
+
+    fn try_map_authorization(&mut self, a: A) -> Result<B, Self::Error> {
+        self.3(a)
+    }
 }
 
 #[derive(Debug, Clone)]